move mesh to vispy
